requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-widget-tweaks==1.4.2
idna==2.6
mysql-connector==2.1.6
mysqlclient==1.3.12
pytz==2018.4
requests==2.18.4
urllib3==1.22
